+++ change-log.md
@@ -1,5 +1,8 @@
 # 🎉 Danho's Genshin Tool v . 2.1.13 🎉
 
+## Added
+* Added Teyvat regions to artifacts - including the filter in SearchableArtifactList.
+
 ## Updates
 * Updated The Doctor (Dottore)'s description since I've now played through the archon quest :D